Modernize the router package.
1. `params` are now available via `ctx.params` and have better types.
2. Upgrade to eslint
3. Remove path-match dependency, which was deprecated and use
   path-to-regexp instead (and has types)

Access to params via `ctx.state.params` is kept for BC, but will be
removed in a future version.
