+++ .eslintrc.json
@@ -1,0 +1,48 @@
+{
+  "env": {
+      "browser": true,
+      "node": true,
+      "es6": true
+  },
+  "extends": [
+      "eslint:recommended",
+      "plugin:@typescript-eslint/eslint-recommended",
+      "plugin:@typescript-eslint/recommended"
+  ],
+  "globals": {
+      "Atomics": "readonly",
+      "SharedArrayBuffer": "readonly"
+  },
+  "parser": "@typescript-eslint/parser",
+  "parserOptions": {
+      "ecmaVersion": 2018,
+      "sourceType": "module"
+  },
+  "plugins": [
+      "@typescript-eslint"
+  ],
+  "rules": {
+      "indent": ["error", 2, {
+        "SwitchCase": 1
+      }],
+      "linebreak-style": ["error", "unix"],
+      "quotes": ["error", "single",
+          {
+            "allowTemplateLiterals": false,
+            "avoidEscape": true
+          }
+      ],
+      "semi": ["error", "always"],
+      "no-trailing-spaces": "error",
+      "eol-last": "error",
+      "@typescript-eslint/no-inferrable-types": "off",
+      "@typescript-eslint/ban-ts-comment": ["error",
+        {
+          "ts-expect-error": "allow-with-description"
+        }
+      ],
+      "@typescript-eslint/no-explicit-any" : 0,
+      "@typescript-eslint/no-unused-vars": ["error"],
+      "no-console": ["error"]
+  }
+}
