+++ Makefile
@@ -1,24 +1,38 @@
-PATH:=./node_modules/.bin:$(PATH)
+SOURCE_FILES:=$(shell find src/ -type f -name '*.ts')
+
+.PHONY:all
+all: build
 
 .PHONY:build
-build:
-	tsc
+build: dist/build
 
 .PHONY:test
 test:
-	nyc mocha
+	node_modules/.bin/nyc node_modules/.bin/mocha
 
 .PHONY:lint
 lint:
-	tslint -p .
+	node_modules/.bin/eslint --quiet 'src/*.ts' 'test/*.ts'
 
 .PHONY:lint-fix
 lint-fix: fix
 
 .PHONY:fix
 fix:
-	tslint -p . --fix
+	node_modules/.bin/eslint --quiet 'src/**/*.ts' 'test/**/*.ts' --fix
 
 .PHONY:watch
 watch:
-	tsc --watch
+	node_modules/.bin/tsc --watch
+
+.PHONY:start
+start: build
+
+.PHONY:clean
+clean:
+	rm -r dist
+
+dist/build: $(SOURCE_FILES)
+	node_modules/.bin/tsc
+	@# Creating a small file to keep track of the last build time
+	touch dist/build
