--- tslint.json
@@ -1,25 +1,0 @@
-{
-    "defaultSeverity": "error",
-    "extends": [
-        "tslint:recommended"
-    ],
-    "jsRules": {},
-    "rules": {
-      "arrow-parens": false,
-      "interface-name": false,
-      "interface-over-type-literal": false,
-      "max-classes-per-file": false,
-      "max-line-length": false,
-      "member-access": false,
-      "member-ordering": false,
-      "no-angle-bracket-type-assertion": false,
-      "no-consecutive-blank-lines": [true, 2],
-      "no-floating-promises": true,
-      "no-var-requires": false,
-      "object-literal-shorthand": false,
-      "object-literal-sort-keys": false,
-      "quotemark": [true, "single", "avoid-template", "avoid-escape"],
-      "trailing-comma": "never"
-    },
-    "rulesDirectory": []
-}
